refactor: encapsulate relay editing and probe validation
